# Performance improvements, refactoring and tidying up readme