--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,7 @@
+module intset
+
+go 1.18
+
+require github.com/karlseguin/expect v1.0.8
+
+require github.com/wsxiaoys/terminal v0.0.0-20160513160801-0940f3fc43a0 // indirect