--- a/license.txt
+++ b/license.txt
@@ -1,3 +1,4 @@
+Copyright (c) 2022 Wu Tingfeng.
 Copyright (c) 2015 Karl Seguin.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy