--- a/sized32.go
+++ b/sized32.go
@@ -109,19 +109,17 @@ func (s Sized32) index(value uint32, bucket []uint32) (int, bool) {
 	if value == v {
 		return l, true
 	}
-
-	var offset, i int
 	if value > v {
-		offset = l
-		bucket = bucket[offset:]
+		bucket = bucket[l:]
 	}
 
+	var i int
 	for i, v = range bucket {
 		if v >= value {
-			return offset + i, v == value
+			return l + i, v == value
 		}
 	}
-	return offset + i + 1, false
+	return l + i + 1, false
 }
 
 func (s Sized32) exists(value uint32, bucket []uint32) bool {