… go.sum
… readme.md
… rune_test.go
… sized32_test.go
… sized_test.go