First, temporarily change the first line in `go.mod` to `module github.com/elliotwutingfeng/intset`

## Changes

- Support rune sets; possible applications include checking strings for prohibited characters etc.
- Improve lookup speed by replacing `index()` method with new `exists()` method that omits unused variables.
- Improve lookup speed by changing bucketSize, bucketGrowBy, bucketMultiplier to 4, 1, 1; tradeoff is **~22%** increase in memory usage compared to existing intset implementation.
- Add docstrings to exported functions.
- README: Fix typos, add badges, mention new rune set feature.
- Add benchmarks for sets implemented with `map[T]struct{}`, for comparison with intset.

## Benchmarks

Memory usage benchmarking done with **pprof**

Speed benchmarking done using vscode's golang codelens extension

Benchmarks performed on AMD Ryzen 7 5800X, Manjaro Linux.

```bash
go test -cpuprofile cpu.prof -memprofile mem.prof -bench . -cpu 1
go tool pprof mem.prof
web
```

| bucketSize                                     | Benchmark_RuneMapDenseExists | Benchmark_RuneMapSparseExists | Total Memory Usage |
|------------------------------------------------|------------------------------|-------------------------------|--------------------|
| 1                                              | 1.215ns/op                   | 1.142ns/op                    | 15916.34MB         |
| 2                                              | 1.348ns/op                   | 1.165ns/op                    | 9075.40MB          |
| 4                                              | 1.638ns/op                   | 1.391ns/op                    | 6961.48MB          |
| 8                                              | 2.156ns/op                   | 1.451ns/op                    | 7268.57MB          |
| 32 (bucketGrowBy = 4 and bucketMultiplier = 2) | 4.002ns/op                   | 1.654ns/op                    | 5713.25MB          |

- The bottom row is the original intset implementation.
- Benchmarks indicate that intset lookups are approximately 7 times faster than traditional sets implemented with `map[T]struct{}`.
- Smaller bucketSize is faster but memory usage is very high.
- Strangely, bucketSize 8 has slightly higher memory usage than bucketSize 4.
- bucketSize 4 was chosen as it gives the best speed-memory tradeoff.