@karlseguin 2022-07-13
Hi.

It's a nice performance improvement. I'm happy to merge this as-is. But after that, what would you think if I made `bucketSize` and `bucketGrowBy` configurable on a per-set basis, so that people could pick the best balance of performance/memory.

Essentially, `Sized`, `Sized32` and `Rune` would gain 1 extra field, `bucketGrowBy`, and new constuctors (or a config struct) would accept a bucketSize and  bucketGrowBy, with the defaults being what you have here.

??